convert Selection to typescript
